move FilesSearch
