+++ src/pages/FilesPage.js
@@ -6,7 +6,7 @@ import AuthService from "../services/AuthService";
 import { FileService } from '../services/FileService';
 import { useParams, useNavigate } from 'react-router-dom';
 import { useLocation } from 'react-router-dom';
-import { FileSearch } from '../components/FileSearch'
+import { FileSearch } from '../components/FilesSearch'
 
 // Импортируем новые компоненты
 import Breadcrumbs from '../components/Breadcrumbs';
@@ -489,130 +489,3 @@ const FilesPage = () => {
 };
 
 export default FilesPage;
-
-
-
-/*
-<div className="container mx-auto px-4 py-8">
-				<div className="mb-6">
-					<h1 className="text-3xl font-bold mb-2">Файловый менеджер</h1>
-					<Breadcrumbs 
-						currentPath={currentPath} 
-						onNavigate={navigateToFolder} 
-					/>
-				</div>
-				
-				<div className="mb-6 bg-gray-50 p-4 rounded-lg">
-					<div className="flex flex-col md:flex-row md:items-center space-y-3 md:space-y-0 md:space-x-4">
-						<div className="flex-1">
-							<form onSubmit={handlePathSubmit} className="flex">
-								<input
-									type="text"
-									value={pathInput}
-									onChange={handlePathInputChange}
-									onKeyDown={handlePathInputKeyDown}
-									placeholder="Введите путь (например: documents/images)"
-								/>
-								<button
-									type="submit"
-								>
-									Перейти
-								</button>
-							</form>
-						</div>
-						
-						<div className="flex flex-wrap gap-2">
-							<button
-								onClick={navigateUp}
-								disabled={!currentPath}
-								className="px-4 py-2 bg-gray-200 rounded hover:bg-gray-300 disabled:opacity-50"
-								type="button"
-							>
-								Назад
-							</button>
-							
-							<button
-								onClick={() => setShowCreateFolderModal(true)}
-								className="px-4 py-2 bg-green-500 text-white rounded hover:bg-green-600"
-								type="button"
-							>
-								Создать папку
-							</button>
-							
-							<label className="px-4 py-2 bg-blue-500 text-white rounded hover:bg-blue-600 cursor-pointer">
-								{uploading ? 'Загрузка...' : 'Загрузить файл'}
-								<input
-									type="file"
-									className="hidden"
-									onChange={handleFileUpload}
-									disabled={uploading}
-								/>
-							</label>
-						</div>
-					</div>
-					
-					<div className="mt-2 text-sm text-gray-600">
-						<span className="font-medium">Текущий путь:</span> {currentPath || '/'}
-					</div>
-				</div>
-				
-				<ErrorMessage 
-					message={error}
-					onClose={() => setError('')}
-					showNavigation={true}
-					onNavigateToRoot={navigateToRoot}
-					onNavigateUp={navigateUp}
-					showUpButton={!!currentPath}
-				/>
-				
-				{loading && (
-					<div className="text-center py-8">
-						<div className="animate-spin rounded-full h-12 w-12 border-b-2 border-blue-500 mx-auto"></div>
-						<p className="mt-2 text-gray-600">Загрузка файлов...</p>
-					</div>
-				)}
-				
-				{!loading && (
-					<>
-						<div className="mb-8">
-							<h2 className="text-xl font-semibold mb-4 flex items-center">
-								<span className="mr-2">📁</span> Папки ({folders.length})
-							</h2>
-							<FolderGrid 
-								folders={folders}
-								onNavigate={navigateToFolder}
-								onDelete={prepareDelete}
-							/>
-						</div>
-						
-						<div>
-							<h2 className="text-xl font-semibold mb-4 flex items-center">
-								<span className="mr-2">📄</span> Файлы ({files.length})
-							</h2>
-							<FileTable 
-								files={files}
-								onDownload={handleDownload}
-								onDelete={prepareDelete}
-							/>
-						</div>
-					</>
-				)}
-				
-				<CreateFolderModal 
-					isOpen={showCreateFolderModal}
-					onClose={() => setShowCreateFolderModal(false)}
-					currentPath={currentPath}
-					onCreate={handleCreateFolder}
-				/>
-				
-				<DeleteConfirmationModal 
-					isOpen={showDeleteModal}
-					onClose={() => {
-						setShowDeleteModal(false);
-						setItemToDelete(null);
-					}}
-					itemName={itemToDelete?.name}
-					onConfirm={handleDelete}
-				/>
-			</div>
-             */
